Fixed DBObject.partialEmit() and other code in the DBStore dumpSchema() chain to properly record supergash and monitor permissions in a dumped schema file.
I may need to look at putting some code into server start up that will
check to make sure that the supergash persona is defined and connected
to the supergash ownergroup.. as it is, if someone takes supergash out
of the supergash ownergroup, supergash has no priviliges.

Made the InvidDBField.test() output a bit more verbose.