Altered the check for the source directory path. Before it was using some arcane sed-foo that I couldn't understand.
I replaced the sed check with a simple call to 'pwd' to get the working
directory of the configure script, which seems to work.

I also altered the check that checks for a valid source directory to 
reflect our new ant build directory layout. I'm not sure why we need
to check for the well-formedness of the source directory, though...

deepak