FEATURE: Added the requisite ARL preamble
FEATURE: Made 'handles' a List instead of a Vector. Everything else is a list,
and we don't really need the synchronization of a Vector anymore.

BUGFIX: Added locking to all of the methods that require it. The List interface
methods that throw UnsupportedOperationException aren't synchronized since,
well, do you really have to encapsulate a throw in a critical section?

FEATURE: Added a method that allows a user to change, on-the-fly, the row
type of the result set (array-based rows or map-based rows).

deepak