Extensive changes to the Ant build file to support generation of crypto keys and certificates for SSL socket use between the Ganymede server and clients.
Also added the ability to sign the client jar files with the server's private key.

Note that none of this is very secure as we do self-signing of our cert, and the password
for the server's private key is incorporated bodily into this Ant build file, but it'll
at least allow a given client jar file to assure itself that it is speaking to the server
that it is intended for.

I'm looking forward to seeing what can be done with the jar file signing.. I'm not using it
for anything yet, myself.

-- Jon