Added code in Ganymede and DBObjectBase to rewrite old class names when processing older ganymede.db files, for compatibility's sake.
Note that this only covers half the problem, as we have also changed (on our own), the gasharl
custom class packages, to arlut.csd.ganymede.gasharl, rather than the generic arlut.csd.ganymede.custom.

Since we can't be guaranteed that we'll know what to do with tasks and classes that aren't defined
in our main Ganymede codebase, we'll have to leave those for an admin to fix up.

-- Jon