Fixed a number of bugs in the client, including the bug that allowed an object to be deleted in the client then viewed in an empty state through the use of the query result table pop-up menu.  Now, all checks against an object being already marked as deleted are done at the proper place, rather than being done in the tree response code.
Also, I fixed bugzilla bug #14, which caused the title of query windows not to be properly
updated when a transaction is committed and the query windows are refreshed, possibly with
a different number of entries.  While I was working on that, I shuffled some code around in
the client for better thread safety when accessing some windowPanel data structures, as well
as the use of a common windowPanel.setWindowTitle() method for picking and maintaining unique
window names.

-- Jon