Some more improvements to the containerPanel layout.  I took the ID hashing out of JLabelPanel, where it didn't really belong.
Now containerPanel defines 3 different categories of component widths, and the
JLabelPanel is free to fill things out to different alignment points on the right
border of the used space.

It seems like it may still use some work, there's at least one case where a
StringSelector is taking double columns when it doesn't have a choice list
presented.

Or something.

-- Jon