Fixed a tricky bug.  The GanymedeSession has an unreferenced() method, which is used to determine when the client has died and dropped its connection to the server.
However, in the case of the xmlclient, this logic doesn't work properly.  The xmlclient doesn't
give two rats asses about maintaining a reference to the GanymedeSession.  All it cares about is
the GanymedeXMLSession which its talks to in order to do its business.  The ClientBase, however,
assumes that the code that uses the ClientBase will need to get access to the GanymedeSession,
so it goes ahead and pulls the session from the server.

This would be fine, since ClientBase persistently retains the reference to the session, but
the xmlclient doesn't even keep a reference to the ClientBase around when it's through logging
in.  Java 5.0 seems to notice that the ClientBase has gone out of reference *much, much faster*
than the older JVM's did, and so when the xmlclient allows the ClientBase to go out of scope,
the client JVM helpfully tells the Ganymede server that the GanymedeSession is unreferenced,
which causes the GanymedeSession to shut everything down and exit 'gracefully'.

In order to fix this, I have made the GanymedeSession's unreferenced method smarter, in coordination
with the GanymedeXMLSession.  Now, the GanymedeSession unreferenced() method will not trigger
a logout so long as the GanymedeXMLSession is still referenced by the client.

Whew.

-- Jon