Reworked the namespace logic so that non-interactive sessions can properly handle namespace allocation transitions, regardless of the order of operations.
That is, when the xmlclient makes edits to an existing Ganymede data store, the user has no way of
controlling the order in which operations are done.  If a user is wanting to swap names between
two objects, the non-interactive nature of the xmlclient defeats him.

With an interactive client, the user could edit object one, set its name to something temporary,
such as 'IGNORETHIS', then go to the second object and give the second object the first object's
name, then go back to the first object and give it the second object's old name, using the
'IGNORETHIS' bit as a temporary value to enable the swap to take place.

This sort of thing can't be expressed with the xmlclient, so now the Namespace management system
will allow non-interactive clients to have two field bindings 'in flight' during the transaction,
so that the transaction can proceed to completion.  If the non-interactive session doesn't get things
cleaned up before the commit() happens (as by attaching a reserved name to a new object but not
detaching it from the object that previously had that name), a CommitNonFatalException is thrown
during commit time, and a list of conflicting values is presented to the client.

-- Jon