Lots of changes to DBEditObject, DBObject, InvidDBField, and GanymedeSession to support more reliable and retrievable labels in xmlclient dump and load.
The big changes revolve around the definition of some explicit rules for telling when
a DBEditObject subclass can be relied upon to return unique generated labels in the
getLabelHook() method.  DBObject and InvidDBField have been regularized so as to more
clearly guarantee the use of consistent labels for referring to objects by id, and the
GanymedeSession queryDispatch() method has been elaborated (yes, even more) so that it
knows about these rules and can do efficient lookup of objects by the fall-back synthetic
labels used by objects without namespace-constrained label fields or guaranteed unique
getLabelHook() methods.

I also tweaked the two object types in the gasharl schema kit that make use of getLabelHook()
so that the xmlclient can properly use their labels (or not, as the case ought to be).

-- Jon