Mega check-in to add support for sync channel XML generation.
This check-in still does not include anything which actually generates output to a sync channel,
but all of the support is there (I believe) for objects and fields to report their delta state,
and the XMLDumpContext has had intelligence added to it so that it can report to objects which
fields they should write out to a sync channel.

This is at 'clean compile' stage, but I haven't done any testing other than reading the code
medium-carefully.

Note as well that I started cascading some GanyPermissionsException throws in here, which might
possibly surprise some plug-in code that's not expecting it.  I'm expecting to need to do
code fixups on the gasharl schema kit and the userKit to compensate.

-- Jon