Put a documentation note about GanyPermissionsException into CHANGES.
I still need to review the client and server code and make sure the client will behave properly
in the face of GanyPermissionsExceptions being thrown.

I may want to make it so that GanyPermissionsExceptions never surface to the client?

I need to think about this a lot more.  It's probably pretty messy and confused at the
moment.

-- Jon