Okay, here's all the scaffolding for the finalization logic to the DBJournal.
All that's left to do is to actually write the routine in DBEditSet to spin over the defined
Sync Channel objects, identify what objects need to be synced to which channels, and to write
the emitXMLDelta() output to those channels.

Plus actually making DBJournal.replay capable of doing that by itself on server reload.

Almost there, stay on target, etc.

-- Jon