Fixed tons and tons of stuff with the sync channel support.
Sync channels are now properly registered on server start up.

Fields are now properly emitted a sync channel has them configured as 'always'
as well as 'when changed'.

Lots of little finagling with the formatting of the xml.

Still need to figure out what to do about lingering transaction records on startup,
still need to track down why the error cascade when the sync channel stuff was
throwing an exception.

-- Jon