Make sure that we catch all possible Throwables in DBEditSet's commit_writeSyncChannels() method, so that we can rethrow as a CommitFatalException.
Also, I've moved the commit_handlePhase2() method around in DBEditSet, so that we
don't call commitPhase2() on DBEditObjects until we know that we have successfully
finalized the transaction to disk.  This way we don't have the DBEditObjects
plugins doing external actions for transactions that might be undone due to a
lack of disk space.

There is still one problem, however.. in that if a transaction is finalized
but the server is then killed before the DBEditObjects can have their
commitPhase2() methods called, those external actions may not be performed.

Still, that's better than having DBEditObjects doing external actions that
are later undone.

-- Jon