Reworked the transaction consistency logic for persisting transactions to the journal, and for handling the loading thereof therefrom.
The old attempt at doing recovery at startup implicitly assumed that no transactions
would progress after the failed transaction was partially scribed into the journal,
but this is tantamount to requiring that the server be shut down on a failure to
persist a transaction.  There's really no good way to live with that.

Instead, I have made it so that the journal has the ability to rewind itself if
a persisted transaction doesn't succeed through to completion in the sync channel
and finalization stages.  This way if someone tries to commit a very large
transaction that won't fit on disk, we should be able to completely undo the
attempted commit and regain all of that disk space, thereby allowing subsequent
smaller transactions to be processed as space allows.

This also simplifies server start up, as there is no need to rejigger things to
allow for sync channel replay.  The journal remains as it was originally intended,
a means of insuring that changes to the ganymede.db file proper are persisted without
having to write the entire ganymede.db file out with each commit.

I'm very confident that this is the right way to go, and this will simplify things
considerably.

This commit also includes some good fixes to the commit sequence, including properly
synchronizing sequential journal operations inside a synchronized block in
DBEditSet.commit_integrateChanges().

Very good stuff, all.. I still need to test all of this on Monday, though, including
running some simulations of out of disk space conditions.

-- Jon