Simple hackery in the DBStore initialization logic to add the new Ganymede 2.0 DBObjectBase and DBObjectBaseField elements to a pre-existing file.
This is *really* ad-hocery.  We really should have some way (an XML definition?) to specify
the type of schema information that must be present for compliance, and which will do
whatever is necessary to see that the server gets its schema in compliance with the
server code's expectations.

This is a very important thing to get done, I think.

-- Jon