Yeah, let's try to apply the namespace constraint too.
I still want to have a general solution for validating and fixing up the schema at
server startup.

The trick is, I can't really do it just with the xml schema element from the
xmlclient spec.  I want to allow the object and field names to be localized with
localization files, which makes it kind of hard to match against my xml schema
element, unless I want to insist that the localization of the schema be done
by producing localized versions of the xml schema definitions.

I could do it by producing a massively huge if-else tree along the lines of the
current verifySchema2_0() method, with some additional logic for checking just
the appropriate pieces of the object and field definitions.

That sort of thing would be big, though.. perhaps even big enough to be worth
breaking out into its own .java file.

Hmm..

-- Jon