Finished localizing DBObjectBaseField.
Simplified getTypeDesc() so that it just returns the basic type and an indication of whether or not it is
an array field, rather than trying to encode all kinds of stuff into the getTypeDesc string.

Made a bunch of errors be IllegalStateException, as they properly should be, rather than
IllegalArgumentExceptions.

Whew.

-- Jon