FEATURE: Support for fullstate syncing of groups. Incremental support is forthcoming!
FEATURE:
mod_group and add_group functions that do exactly what you think they would.

FEATURE:
I added a topological sort function that will take a list of Ganymede group
objects and sort them such that all of a groups dependencies (groups it is
nested inside) appear before the group itself. This is necessary for adding
groups, since we need to make sure that all of this group's "prerequisite"
groups are in place. Otherwise, the operation will fail unnecessarily.