Make the client and console web start files specify full permission access.
I'd rather not do this, but it's the only way to make Foxtrot work for
doing asynchronous GUI operations when using Java Web Start.

My preference would be to have some way of deploying a full permissions
control file with the JNLP file, so that we could get the open socket
permissions and the GUI thread permissions and still be sandboxed for
local file access, etc., but that has sort of thing has to be configured
in a JVM-specific file on the client ahead of time, which is a big pain
for deployment.

As it happens, the previous <j2ee-application-client-permissions/> element
really wasn't very much better at locking things down than <all-permissions/>
is.. but Foxtrot works with <all-permissions/>, at least.

-- Jon