Significant rework of persona/username handling throughout the server, client, and admin console.
The primary impetus for these changes is that users logging in directly with an extended persona
name (i.e., deepak:supergash) wasn't handled correctly in many places.  Case sensitivity issues
with user/persona names are particularly improved, as is the handling of soft timeouts.

As such, this commit fixes some security issues.

This commit also greatly improves the sensibility of the logging messages relating to timeouts
and what-not.

All of login user/persona handling is still too complex, though.  Miserably complex.

-- Jon