Let's make sure we build out oid_to_dn_map before we process any deletions, otherwise we may find we need to do an oid lookup on an oid that was deleted from LDAP by the time we detected the need to consult the oid_to_dn_map at all.
This bit in particular when deleting a user.. this involved deleting both the user's object
and the user's DN from any group objects the user was a member of.  Before this patch,
the maspOpenLDAPSync logic would fail at this.

Also made a few messages more verbose, and changed a variable name or two to make it more
clear what's going on.

I find I'm really not yet comfortable with Python's loosey-goosey non-static typedness.

I'm also not comfortable with generalized except clauses.. I'm finding too many that assume
that the exception raised is due to one circumstance, when another bit of code in the same
block can raise the same exception.

I'd much rather see custom exceptions defined and raised when we know for sure what happened,
and to let the native Python-generated exceptions percolate up on their own if we don't know
for sure what happened.

-- Jon