Made a few modifications to the client's threading, in keeping with recommendations made in the July 27, 2005 edition of the Core Java Technologies Tech Tips newsletter.
The move from SwingUtilities to EventQueue is intended to remove an
unnecessary wrapping layer from the invokeLater and invokeAndWait
calls.  The addition of setPriority(Thread.NORM_PRIORITY) calls is
to ensure that the background operations are not executed at the higher
level of priority automatically inherited from threads started from the
event processing thread.

-- Jon