More cleanup on the pointer dereference section.  I'm also putting in a slightly more informative table heading for the field types column in the field test operator table, and I'm putting in a FIXME for a discussion of the types of fields that Ganymede supports.
Generally, I probably need to put in an extensive section describing the
types of fields supported by Ganymede.  I'm just not sure where to put that,
either in this document or elsewhere.  The xml schema guide discusses these,
but that's in the specific context of definining the Ganymede schema in XML.

-- Jon