Localized PasswordDBField.
This is the last of the DBField subclasses that remained to be localized.  I have simplified a lot
of the messages in this class, to streamline user dialogs.  In many cases I have removed identifying
information about which object and field was being acted upon at the time the error result was returned.

This is fine for interactive usage, but for XML batch mode processing, I'd need to make sure that
some other layer presented the context in which the error occurred.

-- Jon