Mega-wacky patch!
This patch fixes up the gasharl schema's interfaceCustom so that, rather than simply refusing any
mac address submission that doesn't meet our rigorous formatting preferences, it will actually
reformat the submission on the fly and let the client know that it needs to refresh the display
of the just-submitted mac address.

This took a fair bit of reworking of several of the GUI fields from the JDataComponent package,
as well as the logic in the containerPanel pertaining to field refresh.  There were also changes
made to the ReturnVal class so that the verifyNewValue() method in DBEditObject can pass the
need for an auto-refresh back through to the client.

It all compiles cleanly and I've reviewed the patch a few times, but I've not tested anything yet.

-- Jon