Added support to the client for the use of the Java 1.4 Preferences API.
Right now, I'm saving and restoring the size of the main client window,
and making the object creation dialog and the query dialog remember
what object type was last created or searched upon.

This will finally get rid of the annoying 'admin persona' default
for these dialogs.

Now that I understand how all of this works and have written a helper
class for windowSizing, there are plenty more opportunities for incorporating
memory into the client.

For instance, should we remember the position of the login box, or is the
centering behavior we recently implemented enough to please folks?  Should
we implement position remembering for all dialogs?

-- Jon