More work on window position retention.
Added support for position retention for the admin console, including
the vertical position of the splitter pane.. I'm always having to manually
tweak that stuff, sigh.

Also cleaned up and regularized the preference saving for the querybox,
including a modification of the querybox so that it automatically sets
the focus to the object type combo box.

Moved the windowSizer class from arlut.csd.ganymede.client to
arlut.csd.ganymede.common, so that it can be used for the client
and the admin console, both.

-- Jon