Reworked the logic in InvidDBField.bind() so that bind will create the target InvidDBField if it was undefined when the target object was checked out for editing.
This fixes a problem that could block anonymous linking if the user was not explicitly granted
permission to create the invid field by way of a Role object.

Some other small cleanups and code reformatting to try to make at least a little bit easier to read.

Also fixed a typo that left the 'anonymous2' variable in bind unused, and which led to the
linking code at the end of this method to conflate permission to anonymously unlink an old
link and the permission to anonymously link the new link.

-- Jon