Ditched reflection in the DBField static factory methods for performance's sake.
Note that I did not profile any of this, and I suspect the amount of time lost to reflection
overhead would be pretty minimal in most cases, and dominated by RMI overhead in any case, but
in the case of large scale XML operations, this could affect us.

Dropping reflection in these three factory methods means that there are three switches to examine
rather than one, whenever we add a new field type, but they are at least all in one place.

-- Jon