Fixed the client so that it closes view windows on deleted objects at transaction commit time.
Possibly we should want to close _all_ object windows at commit time, but we haven't in the past,
and it might possibly surprise someone.  Probably this is bogus reasoning.

Also modified the server so that it won't throw an exception if the client attempts to view the
object history for a deleted object.  Also modified the server so that supergash accounts can look at
object histories for deleted objects without problems.

-- Jon