Fixed a possible exception in userCustom.consistencyCheck() in the gasharl schema kit.
The fix makes it so that it won't throw an exception if a user object's
group list is null at user consistency check time.

We shouldn't allow a null group list vector in ordinary circumstances, but
it may happen if a user object is having its consistency checked in a
non-transactional context, as with running the database validation task
from the admin console, when an user inactivation is followed by some event
that deletes the user's only group.

The gasharl schema kit should prevent this from occuring due to interactive
use, but the xml client and/or the nightly object inactivation/removal task
might cause this to happen under some circumstances.

We purposely (albeit not wisely?) allow the nightly inactivation/removal
task to leave other objects in the database in an inconsistent state, on
the theory that the task's primary need is to get objects scheduled for
inactivation or removal out of the system.

Such inconsistencies are then to be detected by the database validation
task in the admin console, but that won't work well if an object's
consistencyCheck() method throws an exception.

-- Jon