Cleaned up the new DHCP generation code.
I got rid of the temporary file that was used to prepare the first pass
of host and network definitions in favor of using a StringWriter to produce
the dhcp data in memory during the duration of the writeDHCPFile() method.

Various other cleanups in the DHCP generation code, including the removal
of whitespace before trailing semicolons and the correction of syntax for
the 'allow known clients' blog.

-- Jon