Removed redundant link clearing logic for edit-in-place fields in InvidDBField.
The Ganymede server is designed to automatically clear links pointing to
deleted objects, and we have now resolved the problem that could cause that
mechanism to fail for embedded objects configured in a certain manner.

Thus, there is no longer any need to manually delete elements from an
edit-in-place invid field's values vector.

-- Jon