Moved the attemptAsymBackLinkClear() call to the end of finalizeRemove().
This is necessary because the deletion of embedded objects is counting
on the backlink structures in the DBStore DBLinkTracker to properly clean
up references in the invid fields containing embedded objects.  We were
previously calling attemptAsymBackLinkClear() early on, which obliterated
the back-link tracking structure for the object being deleted, making
it impossible for the deleteObject() call on embedded objects to remove
the relevant invid from the containing invid field.

-- Jon