Fixed owner logging calculation for objects whose ownership is changed by a transaction.
Previously, if an object was edited and the admin doing the editing
changed the owner group of the object, the record for that changes was
sent to the new owners of the object, but not the old ones.

In addition to fixing this, a lot of the calculateOwnerAddresses() method
was poorly constructed, with redundant data structures and unneeded expense
of computation.