Eliminate massive transient class creation for Enums in scheduleHandle.
Previously, TaskType and TaskStatus were inner classes, which
meant that their class definitions contained references to the variables
in the containing scheduleHandle instance.

Now, they are nested classes, which is to say that there is only one
class for each type, and there is only one instance of each type, rather
than an unlimited amount as before.

This change will markedly reduce the working set size of the Ganymede
server and admin console.