Moved TaskType and TaskStatus out of scheduleHandle.
This is to try and avoid the huge number of custom classes that the
JVM creates when TaskType and TaskStatus are inner classes of
scheduleHandle.

Before making this change, the Ganymede server was absolutely
eating memory in the Class store.  Garbage collection cleaned it up,
but it made for an ugly picture during runtime nonetheless.