Fixes a security hole in the PasswordDBField class.
This hole would not allow a user who did not have editing authority over
a given account to mess with the account's password, but if the user was
granted privilege to edit other fields in the account, that user could construct
a custom remote client that would let him set the password despite
lacking privilege to edit that specific field.

With this patch the privileged setAllHashes() method is no longer available
to be called remotely.  The xmlclient can still submit password hashes
to the server over RMI, but the server-side XML processing code will
not permit an escalation of privilege when setting these hashes.