Render the ownerPanel able to show non-editable owner groups permitted by a modification to anonymousLinkOK().
This is necessary if we're going to permit admins to give objects
to owner groups that they don't have full edit access to.

Unfortunately, the need for this suggests that I need to take a look
at the whole objectCache system in the client, and how it tracks what
is or is not editable.  The server should be making the decisions about
what object handles to show, not the client.