** security patch ** Removed getValueString, getEncodingString from remote db_field interface.
These methods made it possible for users remotely logged into the Ganymede
server to access information that was meant to be hidden from them.  In
order for the user to access restricted information, the user would have
to have permission to get a read-only or editable copy of a Ganymede
database object via GanymedeSession.view_db_object() or edit_db_object().

Once they had access to such an object, they were able to call getValueString()
or getEncodingString() on any field in the object, even if those fields
were not meant to be readable by the user.

This did not put password data at risk because PasswordDBFields have never
returned plain or hashtext via either of these methods.  It could reveal
Strings from fields that were not meant to be visible to the user.

In order to have taken advantage of this vulnerability, users would have
had to have built a custom client that abused these methods.  Not a trivial
matter, but not impossible either.