Improved StringSelector's auto-complete logic.
Formerly, the auto-complete logic filled in the rest of the text box
with the matched value and selected the remainder so that further typing
would overwrite it.  However, the cursor was not positioned properly, and
would be left after the auto-complete text selection.

By moving the cursor ourselves before doing the text selection, we are
covered two ways, by having the appropriate text highlighted, and by having
the cursor's position before the highlighted text.