Made the DBPermissionManager refuse to configure itself more than once.
The logic for this is along the lines of making String and Integer final
classes.  It's still possible for custom code running in the Ganymede
server to call the selectPersona() method, but that method at least
requires access to the persona name and password, and is checked against
the user that the DBPermissionManager is associated with.