Fixed a double-dump-lock deadlock race that can block DBWriteLocks.
The problem was due to the SyncRunner's runFullState() method acquiring a
dump lock and then eventually spawning an XMLTransmitter thread which has
its own dump lock.

If a DBWriteLock enters establish between the time that the runFullState()
method's initial dump lock is acquired and fully established and the
XMLTransmitter thread tries to acquire its own, we have a dead lock and
neither the XMLTransmitter nor the transction commit which is acquiring
a DBWriteLock can proceed.

By making this change to DBWriteLock, we prevent a DBWriteLock from proceeding
to put itself on a wait list until there is neither a DBDumpLock on a
waiting list nor established on a base, thus avoiding the deadlock at the
cost of giving DBDumpLocks a bit better chance of getting priority over
DBWriteLocks at submit time.

Note that I'll be looking at fixing the extraneous DBDumpLock in the SyncRunner's
runFullState() call chain as well, but this fix on DBWriteLock prevents
any other double-dump-lock from being able to deadlock the system.