Added some comments in SyncRunner about the double-dump-lock issue.
It turns out that SyncRunner needs the external DBDumpLock in place to
protect the needBuild semaphore, so that we guarantee that we will not miss
issuing a new sync run due to a commit being written into the database
between the time we set needBuild to false and we obtain the DBDumpLock
in the DBSession.getDataXML() call.

Otherwise, we might neglect to do an immediate second sync after completing
our first sync for changes made during this window.