Fixed DBLog.sendSysEventMail() so it won't cc: automated session names
Previously, automated tasks with synthetic session names
that caused DBLogEvents with ccToSelf set to be generated
in a transaction would cause an error to be reported in
the logging system about an unknown recipient to Qstmp.

Now, we only allow ourselves to ccToSelf for the badpass
system event, in which we might reasonably not have a proper
admin invid set in a DBLogEvent but it still would be worth
our while to try to send mail to whatever gibberish the user
entered at login time.