Update the DBObjectBase timestamp more sparingly
Previously, we marked a DBObjectBase as changed even if the only
field changed in objects of that kind in a transaction were the
built-in types, such as Notes, etc.

Now, we'll only consider a DBObjectBase as changed for the purposes
of gating builds on the backend if any of the actual significant
data carrying fields are modified in a transaction.

This will have the effect of reducing unnecessary builds when nothing
but object metadata is changed.

** Note that this will break build logic if anyone was changing builds
based on, e.g., the owner list of an object rather than custom fields.