Revert "Update the DBObjectBase timestamp more sparingly"
This reverts commit 9ba7026c3414dc302bd2c62cd5f3c41e0d6a2f02.

Again, this was unwise.  The DBObjectBase timestamp should still be
updated whenever any metadata is changed on objects of that type.  The
creation date / creation text / last modified date / last modified
text is never going to be the only thing to change an object in any
event, because we don't count objects changed if none of their other
data changed anyway.