Define java.rmi.server.useCodebaseOnly to true in the client launchers.
This configures the client jvms so that they won't allow class files to be
downloaded from the server.  We never need that behavior, so closing this
hole is all to the good.

Oracle made this behavior the default in their Java 1.7.0 release 21
security update, this commit takes care of it when running Ganymede
on older Java runtimes.