Force file sync when writing XML transactions in SyncRunner.
Modified XMLDumpContext and SyncRunner to permit SyncRunner to
do a sync() of the blocks to disk to ensure that the file is completely
available on disk before SyncRunner issues an incremental or
full state sync.

The lack of this sync has not proven to be a problem in the past with
our usage, but sync'ing on our XML transactions should help avoid
any possible races at build time.