[INCOMPATIBLE CHANGE] Conversion of IPDBField from Byte[] to IPAddress.
The new arlut.csd.ganymede.common.IPAddress class encapsulates a Byte[]
so that it cannot be altered after the IPAddress object is created.

All Ganymede code that was previously aware of the Byte[] representation
of IP addresses has been converted to use the new IPAddress class, including
the use of a serializable IPAddress object on the wire for RMI calls.

This change is incompatible with custom GanymedeBuilderTask and DBEditObject
subclasses which handle Byte[] values.  Such code should expect to see
IPAddress objects instead of Byte[] instances when processing Ganymede
objects which contain IP fields.