Taught FieldOptionDBField to avoid doing unnecessary cleaning.
Previously, FieldOptionDBField would be constantly scanning itself
to try to get rid of any references to fields that had been removed
from the Ganymede server with a schema edit.

Now, this is only done if the server's schema has been edited since
server startup and we're doing an emit or we're doing a receive.