FindBugs complained about default character encoding in crypt classes
I'm not sure that using UTF-8 universally is always the right thing, but
it's probably the best thing we can do here.

For most American users, UTF-8 is equivalent to US-ASCII, but users
depending on character encoding matching the host system's
encoding may have problems with this change.  Hopefully everyone's using
UTF-8 on Linux, though.