Rewrote DBPermissionManager.updatePerms() to use DBReadLock for consistency
Previously, there was nothing to stop transactions committing changes on
the roleBase and/or personaBase during updatePerms(), which could lead to
inconsistent permissions being returned from the getPerm() methods in
DBPermissionManager.

This commit also fixes a recent inversion in the rolesWereChanged() and
personaWasChanged() methods which caused updatePerms() to always do a full
recalc of perms, even if the roleBase and personaBase had not been changed
since the last call to updatePerms().